Support running unrun migration queries against live db
