+++ migrator/src/database/mod.rs
@@ -9,6 +9,7 @@ mod embedded_migration;
 mod file_manager;
 mod full_db_info;
 mod left_db;
+mod live_db;
 mod migration_meta;
 mod migrator_db;
 mod queries;
@@ -18,6 +19,7 @@ pub use embedded_migration::*;
 pub use file_manager::*;
 pub use full_db_info::*;
 pub use left_db::*;
+pub use live_db::*;
 pub use migration_meta::*;
 pub use migrator_db::*;
 pub use queries::*;
